Added vue-router and separated demos
