+++ src/main.js
@@ -1,6 +1,20 @@
 import { createApp } from 'vue'
+import { createRouter, createWebHistory } from 'vue-router'
+
 import './style.css'
 import App from './App.vue'
 
-createApp(App)
+import { routes } from './router'
+
+const router = createRouter({
+  history: createWebHistory(),
+  routes
+})
+
+const app = createApp(App)
+
+console.log(routes)
+
+app
+  .use(router)
   .mount('#app')
